On branch main  Your branch is up to date with 'origin/main'.
Spark mlengine should now work, needs it's own sql connector so we don't
have to send json data to it every time (it can fetch from db) and to
implement a collaborative filtering algo based on the data we send.
Collaborative filtering should be easy enough with spark, we just need
to structure data correctly.